move all area endpoints to the area service
